Fixed Water and TMP + Scene Reset
Water does now work with the slider going up when touching an object tagged with "Water", the value does not exceed 100f even after staying there for extended periods of time, and continues to decrease in slider/health when not touching the tagged object. So far only works on the torso/ object attached onto the script with a collider.

The text mesh pro is now fixed showing the value of the slider not showing backwards anymore and value correlates with the value of the slider.

The scene now resets using scene manager in unity when the slider value reaches 0f.