Player Switch + Attempt at NPC/EnemyAI following AI
Player switch causes the octopus to go to the correct position when you press tab on the player.

EnemyAI(4 NPCs when spotting the player separate form the human) does not follow the player correctly and stops after it approaches the player (check and look at what the issue is) + add debugs to know which state the enemy is currently in.

Random movement does work but does cause some issue with the lag of the game when first test.