Player Attaches to Humans to Control Them
The player character can now attach to humans. Doing so will attach the player to the humans head. The player can then control the human to make them move and jump and turn in the direction the player is facing.

TODO:
Add an option to change the players look sensitivity and have that affect the Free Look camera.